fused adetailer with main script
