+++ LightDiffusion.py
@@ -8044,6 +8044,1354 @@ class UltimateSDUpscale:
         return (tensor,)
 
 
+from collections import namedtuple
+import math
+import os
+import re
+import numpy as np
+import torch
+from segment_anything import SamPredictor, sam_model_registry
+
+
+def sam_predict(predictor, points, plabs, bbox, threshold):
+    point_coords = None if not points else np.array(points)
+    point_labels = None if not plabs else np.array(plabs)
+
+    box = np.array([bbox]) if bbox is not None else None
+
+    cur_masks, scores, _ = predictor.predict(
+        point_coords=point_coords, point_labels=point_labels, box=box
+    )
+
+    total_masks = []
+
+    selected = False
+    max_score = 0
+    max_mask = None
+    for idx in range(len(scores)):
+        if scores[idx] > max_score:
+            max_score = scores[idx]
+            max_mask = cur_masks[idx]
+
+        if scores[idx] >= threshold:
+            selected = True
+            total_masks.append(cur_masks[idx])
+        else:
+            pass
+
+    if not selected and max_mask is not None:
+        total_masks.append(max_mask)
+
+    return total_masks
+
+
+def is_same_device(a, b):
+    a_device = torch.device(a) if isinstance(a, str) else a
+    b_device = torch.device(b) if isinstance(b, str) else b
+    return a_device.type == b_device.type and a_device.index == b_device.index
+
+
+class SafeToGPU:
+    def __init__(self, size):
+        self.size = size
+
+    def to_device(self, obj, device):
+        if is_same_device(device, "cpu"):
+            obj.to(device)
+        else:
+            if is_same_device(obj.device, "cpu"):  # cpu to gpu
+                free_memory(self.size * 1.3, device)
+                if get_free_memory(device) > self.size * 1.3:
+                    try:
+                        obj.to(device)
+                    except:
+                        print(
+                            f"WARN: The model is not moved to the '{device}' due to insufficient memory. [1]"
+                        )
+                else:
+                    print(
+                        f"WARN: The model is not moved to the '{device}' due to insufficient memory. [2]"
+                    )
+
+
+class SAMWrapper:
+    def __init__(self, model, is_auto_mode, safe_to_gpu=None):
+        self.model = model
+        self.safe_to_gpu = safe_to_gpu if safe_to_gpu is not None else SafeToGPU_stub()
+        self.is_auto_mode = is_auto_mode
+
+    def prepare_device(self):
+        if self.is_auto_mode:
+            device = get_torch_device()
+            self.safe_to_gpu.to_device(self.model, device=device)
+
+    def release_device(self):
+        if self.is_auto_mode:
+            self.model.to(device="cpu")
+
+    def predict(self, image, points, plabs, bbox, threshold):
+        predictor = SamPredictor(self.model)
+        predictor.set_image(image, "RGB")
+
+        return sam_predict(predictor, points, plabs, bbox, threshold)
+
+
+class SAMLoader:
+    def load_model(self, model_name, device_mode="auto"):
+        modelname = ".\\_internal\\yolos\\" + model_name
+
+        if "vit_h" in model_name:
+            model_kind = "vit_h"
+        elif "vit_l" in model_name:
+            model_kind = "vit_l"
+        else:
+            model_kind = "vit_b"
+
+        sam = sam_model_registry[model_kind](checkpoint=modelname)
+        size = os.path.getsize(modelname)
+        safe_to = SafeToGPU(size)
+
+        # Unless user explicitly wants to use CPU, we use GPU
+        device = get_torch_device() if device_mode == "Prefer GPU" else "CPU"
+
+        if device_mode == "Prefer GPU":
+            safe_to.to_device(sam, device)
+
+        is_auto_mode = device_mode == "AUTO"
+
+        sam_obj = SAMWrapper(sam, is_auto_mode=is_auto_mode, safe_to_gpu=safe_to)
+        sam.sam_wrapper = sam_obj
+
+        print(f"Loads SAM model: {modelname} (device:{device_mode})")
+        return (sam,)
+
+
+from PIL import Image
+import cv2
+import numpy as np
+import torch
+
+orig_torch_load = torch.load
+
+from ultralytics import YOLO
+
+# HOTFIX: https://github.com/ltdrdata/ComfyUI-Impact-Pack/issues/754
+# importing YOLO breaking original torch.load capabilities
+torch.load = orig_torch_load
+
+
+def load_yolo(model_path: str):
+    try:
+        return YOLO(model_path)
+    except ModuleNotFoundError:
+        print("please download yolo model")
+
+
+def inference_bbox(
+    model,
+    image: Image.Image,
+    confidence: float = 0.3,
+    device: str = "",
+):
+    pred = model(image, conf=confidence, device=device)
+
+    bboxes = pred[0].boxes.xyxy.cpu().numpy()
+    cv2_image = np.array(image)
+    cv2_image = cv2_image[:, :, ::-1].copy()  # Convert RGB to BGR for cv2 processing
+    cv2_gray = cv2.cvtColor(cv2_image, cv2.COLOR_BGR2GRAY)
+
+    segms = []
+    for x0, y0, x1, y1 in bboxes:
+        cv2_mask = np.zeros(cv2_gray.shape, np.uint8)
+        cv2.rectangle(cv2_mask, (int(x0), int(y0)), (int(x1), int(y1)), 255, -1)
+        cv2_mask_bool = cv2_mask.astype(bool)
+        segms.append(cv2_mask_bool)
+
+    results = [[], [], [], []]
+    for i in range(len(bboxes)):
+        results[0].append(pred[0].names[int(pred[0].boxes[i].cls.item())])
+        results[1].append(bboxes[i])
+        results[2].append(segms[i])
+        results[3].append(pred[0].boxes[i].conf.cpu().numpy())
+
+    return results
+
+
+def _tensor_check_image(image):
+    return
+
+
+def tensor2pil(image):
+    _tensor_check_image(image)
+    return Image.fromarray(
+        np.clip(255.0 * image.cpu().numpy().squeeze(0), 0, 255).astype(np.uint8)
+    )
+
+
+def create_segmasks(results):
+    bboxs = results[1]
+    segms = results[2]
+    confidence = results[3]
+
+    results = []
+    for i in range(len(segms)):
+        item = (bboxs[i], segms[i].astype(np.float32), confidence[i])
+        results.append(item)
+    return results
+
+
+def dilate_masks(segmasks, dilation_factor, iter=1):
+    dilated_masks = []
+    kernel = np.ones((abs(dilation_factor), abs(dilation_factor)), np.uint8)
+
+    for i in range(len(segmasks)):
+        cv2_mask = segmasks[i][1]
+
+        dilated_mask = cv2.dilate(cv2_mask, kernel, iter)
+
+        item = (segmasks[i][0], dilated_mask, segmasks[i][2])
+        dilated_masks.append(item)
+
+    return dilated_masks
+
+
+def normalize_region(limit, startp, size):
+    if startp < 0:
+        new_endp = min(limit, size)
+        new_startp = 0
+    elif startp + size > limit:
+        new_startp = max(0, limit - size)
+        new_endp = limit
+    else:
+        new_startp = startp
+        new_endp = min(limit, startp + size)
+
+    return int(new_startp), int(new_endp)
+
+
+def make_crop_region(w, h, bbox, crop_factor, crop_min_size=None):
+    x1 = bbox[0]
+    y1 = bbox[1]
+    x2 = bbox[2]
+    y2 = bbox[3]
+
+    bbox_w = x2 - x1
+    bbox_h = y2 - y1
+
+    crop_w = bbox_w * crop_factor
+    crop_h = bbox_h * crop_factor
+
+    kernel_x = x1 + bbox_w / 2
+    kernel_y = y1 + bbox_h / 2
+
+    new_x1 = int(kernel_x - crop_w / 2)
+    new_y1 = int(kernel_y - crop_h / 2)
+
+    # make sure position in (w,h)
+    new_x1, new_x2 = normalize_region(w, new_x1, crop_w)
+    new_y1, new_y2 = normalize_region(h, new_y1, crop_h)
+
+    return [new_x1, new_y1, new_x2, new_y2]
+
+
+def crop_ndarray2(npimg, crop_region):
+    x1 = crop_region[0]
+    y1 = crop_region[1]
+    x2 = crop_region[2]
+    y2 = crop_region[3]
+
+    cropped = npimg[y1:y2, x1:x2]
+
+    return cropped
+
+
+def crop_ndarray4(npimg, crop_region):
+    x1 = crop_region[0]
+    y1 = crop_region[1]
+    x2 = crop_region[2]
+    y2 = crop_region[3]
+
+    cropped = npimg[:, y1:y2, x1:x2, :]
+
+    return cropped
+
+
+crop_tensor4 = crop_ndarray4
+
+
+def crop_image(image, crop_region):
+    return crop_tensor4(image, crop_region)
+
+
+SEG = namedtuple(
+    "SEG",
+    [
+        "cropped_image",
+        "cropped_mask",
+        "confidence",
+        "crop_region",
+        "bbox",
+        "label",
+        "control_net_wrapper",
+    ],
+    defaults=[None],
+)
+
+
+class UltraBBoxDetector:
+    bbox_model = None
+
+    def __init__(self, bbox_model):
+        self.bbox_model = bbox_model
+
+    def detect(
+        self, image, threshold, dilation, crop_factor, drop_size=1, detailer_hook=None
+    ):
+        drop_size = max(drop_size, 1)
+        detected_results = inference_bbox(self.bbox_model, tensor2pil(image), threshold)
+        segmasks = create_segmasks(detected_results)
+
+        if dilation > 0:
+            segmasks = dilate_masks(segmasks, dilation)
+
+        items = []
+        h = image.shape[1]
+        w = image.shape[2]
+
+        for x, label in zip(segmasks, detected_results[0]):
+            item_bbox = x[0]
+            item_mask = x[1]
+
+            y1, x1, y2, x2 = item_bbox
+
+            if (
+                x2 - x1 > drop_size and y2 - y1 > drop_size
+            ):  # minimum dimension must be (2,2) to avoid squeeze issue
+                crop_region = make_crop_region(w, h, item_bbox, crop_factor)
+
+                cropped_image = crop_image(image, crop_region)
+                cropped_mask = crop_ndarray2(item_mask, crop_region)
+                confidence = x[2]
+                # bbox_size = (item_bbox[2]-item_bbox[0],item_bbox[3]-item_bbox[1]) # (w,h)
+
+                item = SEG(
+                    cropped_image,
+                    cropped_mask,
+                    confidence,
+                    crop_region,
+                    item_bbox,
+                    label,
+                    None,
+                )
+
+                items.append(item)
+
+        shape = image.shape[1], image.shape[2]
+        segs = shape, items
+
+        return segs
+
+
+class UltraSegmDetector:
+    bbox_model = None
+
+    def __init__(self, bbox_model):
+        self.bbox_model = bbox_model
+
+
+class NO_SEGM_DETECTOR:
+    pass
+
+
+class UltralyticsDetectorProvider:
+    def doit(self, model_name):
+        model = load_yolo(".\\_internal\\yolos\\" + model_name)
+        return UltraBBoxDetector(model), UltraSegmDetector(model)
+
+
+class SEGSLabelFilter:
+    @staticmethod
+    def filter(segs, labels):
+        labels = set([label.strip() for label in labels])
+        return (
+            segs,
+            (segs[0], []),
+        )
+
+
+class BboxDetectorForEach:
+    def doit(
+        self,
+        bbox_detector,
+        image,
+        threshold,
+        dilation,
+        crop_factor,
+        drop_size,
+        labels=None,
+        detailer_hook=None,
+    ):
+        segs = bbox_detector.detect(
+            image, threshold, dilation, crop_factor, drop_size, detailer_hook
+        )
+
+        if labels is not None and labels != "":
+            labels = labels.split(",")
+            if len(labels) > 0:
+                segs, _ = SEGSLabelFilter.filter(segs, labels)
+
+        return (segs,)
+
+
+def center_of_bbox(bbox):
+    w, h = bbox[2] - bbox[0], bbox[3] - bbox[1]
+    return bbox[0] + w / 2, bbox[1] + h / 2
+
+
+def make_2d_mask(mask):
+    if len(mask.shape) == 4:
+        return mask.squeeze(0).squeeze(0)
+    elif len(mask.shape) == 3:
+        return mask.squeeze(0)
+    return mask
+
+
+def combine_masks2(masks):
+    mask = torch.from_numpy(np.array(masks[0]).astype(np.uint8))
+    return mask
+
+
+def dilate_mask(mask, dilation_factor, iter=1):
+    return make_2d_mask(mask)
+
+
+def make_3d_mask(mask):
+    if len(mask.shape) == 4:
+        return mask.squeeze(0)
+    elif len(mask.shape) == 2:
+        return mask.unsqueeze(0)
+    return mask
+
+
+def make_sam_mask(
+    sam,
+    segs,
+    image,
+    detection_hint,
+    dilation,
+    threshold,
+    bbox_expansion,
+    mask_hint_threshold,
+    mask_hint_use_negative,
+):
+    sam_obj = sam.sam_wrapper
+    sam_obj.prepare_device()
+
+    try:
+        image = np.clip(255.0 * image.cpu().numpy().squeeze(), 0, 255).astype(np.uint8)
+
+        total_masks = []
+        # seg_shape = segs[0]
+        segs = segs[1]
+        for i in range(len(segs)):
+            bbox = segs[i].bbox
+            center = center_of_bbox(bbox)
+            x1 = max(bbox[0] - bbox_expansion, 0)
+            y1 = max(bbox[1] - bbox_expansion, 0)
+            x2 = min(bbox[2] + bbox_expansion, image.shape[1])
+            y2 = min(bbox[3] + bbox_expansion, image.shape[0])
+            dilated_bbox = [x1, y1, x2, y2]
+            points = []
+            plabs = []
+            points.append(center)
+            plabs = [1]  # 1 = foreground point, 0 = background point
+            detected_masks = sam_obj.predict(
+                image, points, plabs, dilated_bbox, threshold
+            )
+            total_masks += detected_masks
+
+        # merge every collected masks
+        mask = combine_masks2(total_masks)
+
+    finally:
+        sam_obj.release_device()
+
+    mask = mask.float()
+    mask = dilate_mask(mask.cpu().numpy(), dilation)
+    mask = torch.from_numpy(mask)
+
+    mask = make_3d_mask(mask)
+    return mask
+
+
+class SAMDetectorCombined:
+    def doit(
+        self,
+        sam_model,
+        segs,
+        image,
+        detection_hint,
+        dilation,
+        threshold,
+        bbox_expansion,
+        mask_hint_threshold,
+        mask_hint_use_negative,
+    ):
+        return (
+            make_sam_mask(
+                sam_model,
+                segs,
+                image,
+                detection_hint,
+                dilation,
+                threshold,
+                bbox_expansion,
+                mask_hint_threshold,
+                mask_hint_use_negative,
+            ),
+        )
+
+
+def segs_bitwise_and_mask(segs, mask):
+    mask = make_2d_mask(mask)
+    items = []
+
+    mask = (mask.cpu().numpy() * 255).astype(np.uint8)
+
+    for seg in segs[1]:
+        cropped_mask = (seg.cropped_mask * 255).astype(np.uint8)
+        crop_region = seg.crop_region
+
+        cropped_mask2 = mask[
+            crop_region[1] : crop_region[3], crop_region[0] : crop_region[2]
+        ]
+
+        new_mask = np.bitwise_and(cropped_mask.astype(np.uint8), cropped_mask2)
+        new_mask = new_mask.astype(np.float32) / 255.0
+
+        item = SEG(
+            seg.cropped_image,
+            new_mask,
+            seg.confidence,
+            seg.crop_region,
+            seg.bbox,
+            seg.label,
+            None,
+        )
+        items.append(item)
+
+    return segs[0], items
+
+
+class SegsBitwiseAndMask:
+    def doit(self, segs, mask):
+        return (segs_bitwise_and_mask(segs, mask),)
+
+
+def general_tensor_resize(image, w: int, h: int):
+    _tensor_check_image(image)
+    image = image.permute(0, 3, 1, 2)
+    image = torch.nn.functional.interpolate(image, size=(h, w), mode="bilinear")
+    image = image.permute(0, 2, 3, 1)
+    return image
+
+
+class TensorBatchBuilder:
+    def __init__(self):
+        self.tensor = None
+
+    def concat(self, new_tensor):
+        self.tensor = new_tensor
+
+
+def pil2tensor(image):
+    return torch.from_numpy(np.array(image).astype(np.float32) / 255.0).unsqueeze(0)
+
+
+LANCZOS = Image.Resampling.LANCZOS if hasattr(Image, "Resampling") else Image.LANCZOS
+
+
+def tensor_resize(image, w: int, h: int):
+    _tensor_check_image(image)
+    if image.shape[3] >= 3:
+        scaled_images = TensorBatchBuilder()
+        for single_image in image:
+            single_image = single_image.unsqueeze(0)
+            single_pil = tensor2pil(single_image)
+            scaled_pil = single_pil.resize((w, h), resample=LANCZOS)
+
+            single_image = pil2tensor(scaled_pil)
+            scaled_images.concat(single_image)
+
+        return scaled_images.tensor
+    else:
+        return general_tensor_resize(image, w, h)
+
+
+def segs_scale_match(segs, target_shape):
+    h = segs[0][0]
+    w = segs[0][1]
+
+    th = target_shape[1]
+    tw = target_shape[2]
+
+    if (h == th and w == tw) or h == 0 or w == 0:
+        return segs
+
+
+def starts_with_regex(pattern, text):
+    regex = re.compile(pattern)
+    return regex.match(text)
+
+
+class WildcardChooser:
+    def __init__(self, items, randomize_when_exhaust):
+        self.i = 0
+        self.items = items
+        self.randomize_when_exhaust = randomize_when_exhaust
+
+    def get(self, seg):
+        item = self.items[self.i]
+        self.i += 1
+
+        return item
+
+
+def process_wildcard_for_segs(wildcard):
+    return None, WildcardChooser([(None, wildcard)], False)
+
+
+class DifferentialDiffusion:
+    def apply(self, model):
+        model = model.clone()
+        model.set_model_denoise_mask_function(self.forward)
+        return (model,)
+
+    def forward(
+        self, sigma: torch.Tensor, denoise_mask: torch.Tensor, extra_options: dict
+    ):
+        model = extra_options["model"]
+        step_sigmas = extra_options["sigmas"]
+        sigma_to = model.inner_model.model_sampling.sigma_min
+        sigma_from = step_sigmas[0]
+
+        ts_from = model.inner_model.model_sampling.timestep(sigma_from)
+        ts_to = model.inner_model.model_sampling.timestep(sigma_to)
+        current_ts = model.inner_model.model_sampling.timestep(sigma[0])
+
+        threshold = (current_ts - ts_to) / (ts_from - ts_to)
+
+        return (denoise_mask >= threshold).to(denoise_mask.dtype)
+
+
+def to_tensor(image):
+    return torch.from_numpy(image)
+
+
+import torchvision
+
+
+def _tensor_check_mask(mask):
+    return
+
+
+def tensor_gaussian_blur_mask(mask, kernel_size, sigma=10.0):
+    """Return NHWC torch.Tenser from ndim == 2 or 4 `np.ndarray` or `torch.Tensor`"""
+    if isinstance(mask, np.ndarray):
+        mask = torch.from_numpy(mask)
+
+    if mask.ndim == 2:
+        mask = mask[None, ..., None]
+
+    _tensor_check_mask(mask)
+
+    kernel_size = kernel_size * 2 + 1
+
+    prev_device = mask.device
+    device = get_torch_device()
+    mask.to(device)
+
+    # apply gaussian blur
+    mask = mask[:, None, ..., 0]
+    blurred_mask = torchvision.transforms.GaussianBlur(
+        kernel_size=kernel_size, sigma=sigma
+    )(mask)
+    blurred_mask = blurred_mask[:, 0, ..., None]
+
+    blurred_mask.to(prev_device)
+
+    return blurred_mask
+
+
+def to_latent_image(pixels, vae):
+    x = pixels.shape[1]
+    y = pixels.shape[2]
+    return VAEEncode().encode(vae, pixels)[0]
+
+
+def calculate_sigmas2(model, sampler, scheduler, steps):
+    return calculate_sigmas(
+        model.get_model_object("model_sampling"), scheduler, steps
+    )
+
+
+def get_noise_sampler(x, cpu, total_sigmas, **kwargs):
+    if "extra_args" in kwargs and "seed" in kwargs["extra_args"]:
+        sigma_min, sigma_max = total_sigmas[total_sigmas > 0].min(), total_sigmas.max()
+        seed = kwargs["extra_args"].get("seed", None)
+        return BrownianTreeNoiseSampler(x, sigma_min, sigma_max, seed=seed, cpu=cpu)
+    return None
+
+
+def ksampler2(sampler_name, total_sigmas, extra_options={}, inpaint_options={}):
+    if sampler_name == "dpmpp_2m_sde":
+
+        def sample_dpmpp_sde(model, x, sigmas, **kwargs):
+            noise_sampler = get_noise_sampler(x, True, total_sigmas, **kwargs)
+            if noise_sampler is not None:
+                kwargs["noise_sampler"] = noise_sampler
+
+            return sample_dpmpp_2m_sde(model, x, sigmas, **kwargs)
+
+        sampler_function = sample_dpmpp_sde
+
+    else:
+        return sampler_object(sampler_name)
+
+    return KSAMPLER(sampler_function, extra_options, inpaint_options)
+
+
+class Noise_RandomNoise:
+    def __init__(self, seed):
+        self.seed = seed
+
+    def generate_noise(self, input_latent):
+        latent_image = input_latent["samples"]
+        batch_inds = (
+            input_latent["batch_index"] if "batch_index" in input_latent else None
+        )
+        return prepare_noise(latent_image, self.seed, batch_inds)
+
+
+def sample_with_custom_noise(
+    model,
+    add_noise,
+    noise_seed,
+    cfg,
+    positive,
+    negative,
+    sampler,
+    sigmas,
+    latent_image,
+    noise=None,
+    callback=None,
+):
+    latent = latent_image
+    latent_image = latent["samples"]
+
+    out = latent.copy()
+    out["samples"] = latent_image
+
+    if noise is None:
+        noise = Noise_RandomNoise(noise_seed).generate_noise(out)
+
+    noise_mask = None
+    if "noise_mask" in latent:
+        noise_mask = latent["noise_mask"]
+
+    x0_output = {}
+
+    disable_pbar = not PROGRESS_BAR_ENABLED
+
+    device = get_torch_device()
+
+    noise = noise.to(device)
+    latent_image = latent_image.to(device)
+    if noise_mask is not None:
+        noise_mask = noise_mask.to(device)
+
+    samples = sample_custom(
+        model,
+        noise,
+        cfg,
+        sampler,
+        sigmas,
+        positive,
+        negative,
+        latent_image,
+        noise_mask=noise_mask,
+        disable_pbar=disable_pbar,
+        seed=noise_seed,
+    )
+
+    samples = samples.to(intermediate_device())
+
+    out["samples"] = samples
+    out_denoised = out
+    return out, out_denoised
+
+
+def separated_sample(
+    model,
+    add_noise,
+    seed,
+    steps,
+    cfg,
+    sampler_name,
+    scheduler,
+    positive,
+    negative,
+    latent_image,
+    start_at_step,
+    end_at_step,
+    return_with_leftover_noise,
+    sigma_ratio=1.0,
+    sampler_opt=None,
+    noise=None,
+    callback=None,
+    scheduler_func=None,
+):
+
+    total_sigmas = calculate_sigmas2(model, sampler_name, scheduler, steps)
+
+    sigmas = total_sigmas
+
+    if start_at_step is not None:
+        sigmas = sigmas[start_at_step:] * sigma_ratio
+
+    impact_sampler = ksampler2(sampler_name, total_sigmas)
+
+    res = sample_with_custom_noise(
+        model,
+        add_noise,
+        seed,
+        cfg,
+        positive,
+        negative,
+        impact_sampler,
+        sigmas,
+        latent_image,
+        noise=noise,
+        callback=callback,
+    )
+
+    return res[1]
+
+
+def ksampler_wrapper(
+    model,
+    seed,
+    steps,
+    cfg,
+    sampler_name,
+    scheduler,
+    positive,
+    negative,
+    latent_image,
+    denoise,
+    refiner_ratio=None,
+    refiner_model=None,
+    refiner_clip=None,
+    refiner_positive=None,
+    refiner_negative=None,
+    sigma_factor=1.0,
+    noise=None,
+    scheduler_func=None,
+):
+
+    # Use separated_sample instead of KSampler for `AYS scheduler`
+    # refined_latent = nodes.KSampler().sample(model, seed, steps, cfg, sampler_name, scheduler, positive, negative, latent_image, denoise * sigma_factor)[0]
+    advanced_steps = math.floor(steps / denoise)
+    start_at_step = advanced_steps - steps
+    end_at_step = start_at_step + steps
+    refined_latent = separated_sample(
+        model,
+        True,
+        seed,
+        advanced_steps,
+        cfg,
+        sampler_name,
+        scheduler,
+        positive,
+        negative,
+        latent_image,
+        start_at_step,
+        end_at_step,
+        False,
+        sigma_ratio=sigma_factor,
+        noise=noise,
+        scheduler_func=scheduler_func,
+    )
+
+    return refined_latent
+
+
+def enhance_detail(
+    image,
+    model,
+    clip,
+    vae,
+    guide_size,
+    guide_size_for_bbox,
+    max_size,
+    bbox,
+    seed,
+    steps,
+    cfg,
+    sampler_name,
+    scheduler,
+    positive,
+    negative,
+    denoise,
+    noise_mask,
+    force_inpaint,
+    wildcard_opt=None,
+    wildcard_opt_concat_mode=None,
+    detailer_hook=None,
+    refiner_ratio=None,
+    refiner_model=None,
+    refiner_clip=None,
+    refiner_positive=None,
+    refiner_negative=None,
+    control_net_wrapper=None,
+    cycle=1,
+    inpaint_model=False,
+    noise_mask_feather=0,
+    scheduler_func=None,
+):
+
+    if noise_mask is not None:
+        noise_mask = tensor_gaussian_blur_mask(noise_mask, noise_mask_feather)
+        noise_mask = noise_mask.squeeze(3)
+
+    h = image.shape[1]
+    w = image.shape[2]
+
+    bbox_h = bbox[3] - bbox[1]
+    bbox_w = bbox[2] - bbox[0]
+
+    # for cropped_size
+    upscale = guide_size / min(w, h)
+
+    new_w = int(w * upscale)
+    new_h = int(h * upscale)
+
+    if new_w > max_size or new_h > max_size:
+        upscale *= max_size / max(new_w, new_h)
+        new_w = int(w * upscale)
+        new_h = int(h * upscale)
+
+    if upscale <= 1.0 or new_w == 0 or new_h == 0:
+        print(f"Detailer: force inpaint")
+        upscale = 1.0
+        new_w = w
+        new_h = h
+
+    print(
+        f"Detailer: segment upscale for ({bbox_w, bbox_h}) | crop region {w, h} x {upscale} -> {new_w, new_h}"
+    )
+
+    # upscale
+    upscaled_image = tensor_resize(image, new_w, new_h)
+
+    cnet_pils = None
+
+    # prepare mask
+    latent_image = to_latent_image(upscaled_image, vae)
+    if noise_mask is not None:
+        latent_image["noise_mask"] = noise_mask
+
+    refined_latent = latent_image
+
+    # ksampler
+    for i in range(0, cycle):
+        (
+            model2,
+            seed2,
+            steps2,
+            cfg2,
+            sampler_name2,
+            scheduler2,
+            positive2,
+            negative2,
+            upscaled_latent2,
+            denoise2,
+        ) = (
+            model,
+            seed + i,
+            steps,
+            cfg,
+            sampler_name,
+            scheduler,
+            positive,
+            negative,
+            latent_image,
+            denoise,
+        )
+        noise = None
+
+        refined_latent = ksampler_wrapper(
+            model2,
+            seed2,
+            steps2,
+            cfg2,
+            sampler_name2,
+            scheduler2,
+            positive2,
+            negative2,
+            refined_latent,
+            denoise2,
+            refiner_ratio,
+            refiner_model,
+            refiner_clip,
+            refiner_positive,
+            refiner_negative,
+            noise=noise,
+            scheduler_func=scheduler_func,
+        )
+
+    # non-latent downscale - latent downscale cause bad quality
+    try:
+        # try to decode image normally
+        refined_image = vae.decode(refined_latent["samples"])
+    except Exception as e:
+        # usually an out-of-memory exception from the decode, so try a tiled approach
+        refined_image = vae.decode_tiled(
+            refined_latent["samples"],
+            tile_x=64,
+            tile_y=64,
+        )
+
+    # downscale
+    refined_image = tensor_resize(refined_image, w, h)
+
+    # prevent mixing of device
+    refined_image = refined_image.cpu()
+
+    # don't convert to latent - latent break image
+    # preserving pil is much better
+    return refined_image, cnet_pils
+
+
+def tensor_paste(image1, image2, left_top, mask):
+    """Mask and image2 has to be the same size"""
+    _tensor_check_image(image1)
+    _tensor_check_image(image2)
+    _tensor_check_mask(mask)
+
+    x, y = left_top
+    _, h1, w1, _ = image1.shape
+    _, h2, w2, _ = image2.shape
+
+    # calculate image patch size
+    w = min(w1, x + w2) - x
+    h = min(h1, y + h2) - y
+
+    mask = mask[:, :h, :w, :]
+    image1[:, y : y + h, x : x + w, :] = (1 - mask) * image1[
+        :, y : y + h, x : x + w, :
+    ] + mask * image2[:, :h, :w, :]
+    return
+
+
+def tensor_convert_rgba(image, prefer_copy=True):
+    """Assumes NHWC format tensor with 1, 3 or 4 channels."""
+    _tensor_check_image(image)
+    alpha = torch.ones((*image.shape[:-1], 1))
+    return torch.cat((image, alpha), axis=-1)
+
+
+def tensor_convert_rgb(image, prefer_copy=True):
+    """Assumes NHWC format tensor with 1, 3 or 4 channels."""
+    _tensor_check_image(image)
+    return image
+
+
+def tensor_get_size(image):
+    """Mimicking `PIL.Image.size`"""
+    _tensor_check_image(image)
+    _, h, w, _ = image.shape
+    return (w, h)
+
+
+def tensor_putalpha(image, mask):
+    _tensor_check_image(image)
+    _tensor_check_mask(mask)
+    image[..., -1] = mask[..., 0]
+
+
+class DetailerForEach:
+    @staticmethod
+    def do_detail(
+        image,
+        segs,
+        model,
+        clip,
+        vae,
+        guide_size,
+        guide_size_for_bbox,
+        max_size,
+        seed,
+        steps,
+        cfg,
+        sampler_name,
+        scheduler,
+        positive,
+        negative,
+        denoise,
+        feather,
+        noise_mask,
+        force_inpaint,
+        wildcard_opt=None,
+        detailer_hook=None,
+        refiner_ratio=None,
+        refiner_model=None,
+        refiner_clip=None,
+        refiner_positive=None,
+        refiner_negative=None,
+        cycle=1,
+        inpaint_model=False,
+        noise_mask_feather=0,
+        scheduler_func_opt=None,
+    ):
+
+        image = image.clone()
+        enhanced_alpha_list = []
+        enhanced_list = []
+        cropped_list = []
+        cnet_pil_list = []
+
+        segs = segs_scale_match(segs, image.shape)
+        new_segs = []
+
+        wildcard_concat_mode = None
+        wmode, wildcard_chooser = process_wildcard_for_segs(wildcard_opt)
+
+        ordered_segs = segs[1]
+
+        if (
+            noise_mask_feather > 0
+            and "denoise_mask_function" not in model.model_options
+        ):
+            model = DifferentialDiffusion().apply(model)[0]
+
+        for i, seg in enumerate(ordered_segs):
+            cropped_image = crop_ndarray4(
+                image.cpu().numpy(), seg.crop_region
+            )  # Never use seg.cropped_image to handle overlapping area
+            cropped_image = to_tensor(cropped_image)
+            mask = to_tensor(seg.cropped_mask)
+            mask = tensor_gaussian_blur_mask(mask, feather)
+
+            is_mask_all_zeros = (seg.cropped_mask == 0).all().item()
+            if is_mask_all_zeros:
+                print(f"Detailer: segment skip [empty mask]")
+                continue
+
+            cropped_mask = seg.cropped_mask
+
+            seg_seed, wildcard_item = wildcard_chooser.get(seg)
+
+            seg_seed = seed + i if seg_seed is None else seg_seed
+
+            cropped_positive = [
+                [
+                    condition,
+                    {
+                        k: (
+                            crop_condition_mask(v, image, seg.crop_region)
+                            if k == "mask"
+                            else v
+                        )
+                        for k, v in details.items()
+                    },
+                ]
+                for condition, details in positive
+            ]
+
+            cropped_negative = [
+                [
+                    condition,
+                    {
+                        k: (
+                            crop_condition_mask(v, image, seg.crop_region)
+                            if k == "mask"
+                            else v
+                        )
+                        for k, v in details.items()
+                    },
+                ]
+                for condition, details in negative
+            ]
+
+            orig_cropped_image = cropped_image.clone()
+            enhanced_image, cnet_pils = enhance_detail(
+                cropped_image,
+                model,
+                clip,
+                vae,
+                guide_size,
+                guide_size_for_bbox,
+                max_size,
+                seg.bbox,
+                seg_seed,
+                steps,
+                cfg,
+                sampler_name,
+                scheduler,
+                cropped_positive,
+                cropped_negative,
+                denoise,
+                cropped_mask,
+                force_inpaint,
+                wildcard_opt=wildcard_item,
+                wildcard_opt_concat_mode=wildcard_concat_mode,
+                detailer_hook=detailer_hook,
+                refiner_ratio=refiner_ratio,
+                refiner_model=refiner_model,
+                refiner_clip=refiner_clip,
+                refiner_positive=refiner_positive,
+                refiner_negative=refiner_negative,
+                control_net_wrapper=seg.control_net_wrapper,
+                cycle=cycle,
+                inpaint_model=inpaint_model,
+                noise_mask_feather=noise_mask_feather,
+                scheduler_func=scheduler_func_opt,
+            )
+
+            if not (enhanced_image is None):
+                # don't latent composite-> converting to latent caused poor quality
+                # use image paste
+                image = image.cpu()
+                enhanced_image = enhanced_image.cpu()
+                tensor_paste(
+                    image,
+                    enhanced_image,
+                    (seg.crop_region[0], seg.crop_region[1]),
+                    mask,
+                )  # this code affecting to `cropped_image`.
+                enhanced_list.append(enhanced_image)
+
+            # Convert enhanced_pil_alpha to RGBA mode
+            enhanced_image_alpha = tensor_convert_rgba(enhanced_image)
+            new_seg_image = (
+                enhanced_image.numpy()
+            )  # alpha should not be applied to seg_image
+            # Apply the mask
+            mask = tensor_resize(mask, *tensor_get_size(enhanced_image))
+            tensor_putalpha(enhanced_image_alpha, mask)
+            enhanced_alpha_list.append(enhanced_image_alpha)
+
+            cropped_list.append(orig_cropped_image)  # NOTE: Don't use `cropped_image`
+
+            new_seg = SEG(
+                new_seg_image,
+                seg.cropped_mask,
+                seg.confidence,
+                seg.crop_region,
+                seg.bbox,
+                seg.label,
+                seg.control_net_wrapper,
+            )
+            new_segs.append(new_seg)
+
+        image_tensor = tensor_convert_rgb(image)
+
+        cropped_list.sort(key=lambda x: x.shape, reverse=True)
+        enhanced_list.sort(key=lambda x: x.shape, reverse=True)
+        enhanced_alpha_list.sort(key=lambda x: x.shape, reverse=True)
+
+        return (
+            image_tensor,
+            cropped_list,
+            enhanced_list,
+            enhanced_alpha_list,
+            cnet_pil_list,
+            (segs[0], new_segs),
+        )
+
+
+def empty_pil_tensor(w=64, h=64):
+    return torch.zeros((1, h, w, 3), dtype=torch.float32)
+
+
+class DetailerForEachTest(DetailerForEach):
+    def doit(
+        self,
+        image,
+        segs,
+        model,
+        clip,
+        vae,
+        guide_size,
+        guide_size_for,
+        max_size,
+        seed,
+        steps,
+        cfg,
+        sampler_name,
+        scheduler,
+        positive,
+        negative,
+        denoise,
+        feather,
+        noise_mask,
+        force_inpaint,
+        wildcard,
+        detailer_hook=None,
+        cycle=1,
+        inpaint_model=False,
+        noise_mask_feather=0,
+        scheduler_func_opt=None,
+    ):
+
+        (
+            enhanced_img,
+            cropped,
+            cropped_enhanced,
+            cropped_enhanced_alpha,
+            cnet_pil_list,
+            new_segs,
+        ) = DetailerForEach.do_detail(
+            image,
+            segs,
+            model,
+            clip,
+            vae,
+            guide_size,
+            guide_size_for,
+            max_size,
+            seed,
+            steps,
+            cfg,
+            sampler_name,
+            scheduler,
+            positive,
+            negative,
+            denoise,
+            feather,
+            noise_mask,
+            force_inpaint,
+            wildcard,
+            detailer_hook,
+            cycle=cycle,
+            inpaint_model=inpaint_model,
+            noise_mask_feather=noise_mask_feather,
+            scheduler_func_opt=scheduler_func_opt,
+        )
+
+        cnet_pil_list = [empty_pil_tensor()]
+
+        return (
+            enhanced_img,
+            cropped,
+            cropped_enhanced,
+            cropped_enhanced_alpha,
+            cnet_pil_list,
+        )
+
+
+
 import contextlib
 import functools
 import logging
@@ -8511,7 +9859,6 @@ if glob.glob(".\\_internal\\embeddings\\*.pt") == None :
     hf_hub_download(repo_id="gsdf/EasyNegative", filename="EasyNegative.pt")
 
 
-import wkms
 class App(tk.Tk):
     def __init__(self):
         super().__init__()
@@ -8873,7 +10220,7 @@ class App(tk.Tk):
             except:
                 loraloader_274 = checkpointloadersimple_241
             
-            samloader = wkms.SAMLoader()
+            samloader = SAMLoader()
             samloader_87 = samloader.load_model(
                 model_name="sam_vit_b_01ec64.pth", device_mode="AUTO"
             )
@@ -8883,17 +10230,17 @@ class App(tk.Tk):
                 clip=loraloader_274[1],
             )
 
-            ultralyticsdetectorprovider = wkms.UltralyticsDetectorProvider()
+            ultralyticsdetectorprovider = UltralyticsDetectorProvider()
             ultralyticsdetectorprovider_151 = ultralyticsdetectorprovider.doit(
                 #model_name="face_yolov8m.pt"
                 model_name="person_yolov8m-seg.pt"
             )
 
         
-            bboxdetectorsegs = wkms.BboxDetectorForEach()
-            samdetectorcombined = wkms.SAMDetectorCombined()
-            impactsegsandmask = wkms.SegsBitwiseAndMask()
-            detailerforeachdebug = wkms.DetailerForEachTest()
+            bboxdetectorsegs = BboxDetectorForEach()
+            samdetectorcombined = SAMDetectorCombined()
+            impactsegsandmask = SegsBitwiseAndMask()
+            detailerforeachdebug = DetailerForEachTest()
 
             clipsetlastlayer = CLIPSetLastLayer()
             clipsetlastlayer_257 = clipsetlastlayer.set_last_layer(
@@ -9023,7 +10370,7 @@ class App(tk.Tk):
                     images=detailerforeachdebug_145[0],
                 )
 
-                ultralyticsdetectorprovider = wkms.UltralyticsDetectorProvider()
+                ultralyticsdetectorprovider = UltralyticsDetectorProvider()
                 ultralyticsdetectorprovider_151 = ultralyticsdetectorprovider.doit(
                     model_name="face_yolov9c.pt"
                 )
