+++ _internal/prompt.txt
@@ -1,5 +1,5 @@
-prompt: a picture of a landscape with a castle on top of a mountain
+prompt: a girl with white long hair, sitting in wariza, in a seducing way
 neg: (worst quality, low quality:1.4), (zombie, sketch, interlocked fingers, comic), (embedding:EasyNegative), (embedding:badhandv4)
-w: 768
-h: 512
+w: 512
+h: 1024
 cfg: 8
